% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adonis.comp.R
\name{adonis.comp}
\alias{adonis.comp}
\title{Compute F score.}
\usage{
adonis.comp(tre.mt, groups, permutations = 99, f.perms = FALSE,
  svQTL = FALSE, approx = TRUE)
}
\arguments{
\item{groups}{a factor with the group information.}

\item{permutations}{the number of permutations.}

\item{f.perms}{should the permuted F scores be returned instead of the
real F score. Default is FALSE.}

\item{svQTL}{should svQTL test be performed instead of sQTL. Default is FALSE.}

\item{approx}{should the asymptotic distribution be used instead of permutations.
Default is TRUE.}

\item{dis}{the distance object of the transcript relative expression.}
}
\value{
a vector with the F score or the permuted F scores.
}
\description{
Compute either observed F score or permuted F scores.
}
\author{
Jean Monlong, Diego Garrido-Martín
}
\keyword{internal}
