% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqtl.seeker.R
\name{check.genotype}
\alias{check.genotype}
\title{Check genotype}
\usage{
check.genotype(geno.df, tre.df, min.nb.ind.geno = 10)
}
\arguments{
\item{geno.df}{a data.frame of genotypes produced by \code{\link{read.bedix}}.}

\item{tre.df}{a data.frame with transcript relative expression values
produced by \code{\link{prepare.trans.exp}} corresponding to one gene.}
}
\value{
A character vector of SNP suitabilities. Possible labels:
\item{"Missing genotype" (more than 3 missing genotype values)}{ }
\item{"Only one genotype group"}{ }
\item{"Not all the groups with >10 samples"}{ }
\item{"Not all the groups with >5 different splicing patterns"}{ }
\item{"PASS"}{ }
}
\description{
Labels SNPs according to their suitability for sQTL mapping.
}
\author{
Jean Monlong, Diego Garrido-Martín
}
\keyword{internal}
