% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compPvalue.R
\name{compPvalue}
\alias{compPvalue}
\title{P-values computation}
\usage{
compPvalue(res.df, tre.mt, min.nb.ext.scores = 1000, nb.perm.max = 1e+06,
  svQTL = FALSE, approx = TRUE)
}
\arguments{
\item{res.df}{a data.frame with the F scores and number of groups.}

\item{tre.mt}{a matrix with the transcript relative expression (samples x transcripts).}

\item{min.nb.ext.scores}{the minimum number of permuted scores higher than
'F.lead' to allow the computation to stop. Default is 1000.}

\item{nb.perm.max}{the maximum number of permutations. Default is 1e6.}

\item{svQTL}{should svQTL test be performed instead of sQTL. Default is FALSE.}

\item{approx}{should the asymptotic distribution (Monte Carlo) be used instead of permutations.
Default is TRUE.}
}
\value{
An updated data.frame res.df with new columns 'pv' and 'nb.perms'.
}
\description{
Compute P-values from F scores.
}
\author{
Jean Monlong, Diego Garrido-Martín
}
\keyword{internal}
