% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqtl.seeker.p.R
\name{compute.empirical.pv}
\alias{compute.empirical.pv}
\title{Compute empirical P-value}
\usage{
compute.empirical.pv(genotype.gene, tre.dist, best.snp, min.pv.obs,
  min.nb.ext.scores = 100, nb.perm.max = 1000, comp.ld = TRUE,
  verbose = FALSE)
}
\arguments{
\item{genotype.gene}{a data.frame of genotypes produced by \code{\link{read.bedix}}.}

\item{tre.dist}{a interdistance matrix produced by \code{\link{hellingerDist}}.}

\item{best.snp}{SNP with the smallest observed nominal P-value, computed by \code{\link{compute.nominal.pv}}.}

\item{min.pv.obs}{smallest observed nominal P-value.}

\item{min.nb.ext.scores}{the minimum number of permuted  nominal P-values lower than
the smallest observed nominal P-value to allow the computation to stop. Default is 100.}

\item{nb.perm.max}{the maximum number of permutations. Default is 1000.}

\item{verbose}{Default is TRUE. Mainly for debugging.}
}
\value{
a data.frame with columns:
\item{variants.cis}{the number of variants tested in cis.}
\item{LD}{a linkage disequilibrium estimate for the genomic window (median r2).}
\item{best.snp}{ID of the SNP with the smallest observed nominal P-value.}
\item{best.nominal.pv}{P-value corresponding to the best SNP.}
\item{shape1}{Beta distribution parameter shape1.}
\item{shape2}{Beta distribution parameter shape2.}
\item{nb.perms}{the number of permutations used for the empirical P-value computation.}
\item{pv.emp}{empirical P-value based on permutations.}
\item{pv.emp.beta}{empirical P-value based on the beta approximation.}
\item{runtime}{approximated computation time per gene.}
}
\description{
Computes empirical P-value for a gene.
}
\author{
Diego Garrido-Martín
}
\keyword{internal}
