% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqtl.seeker.p.R
\name{compute.nominal.pv}
\alias{compute.nominal.pv}
\title{Compute nominal P-value}
\usage{
\method{compute}{nominal.pv}(geno.df, tre.dist, permute = FALSE)
}
\arguments{
\item{geno.df}{a data.frame of a single row corresponding to the genotype of one SNP 
observed in several individuals (columns).}

\item{tre.dist}{a interdistance matrix produced by \code{\link{hellingerDist}}.}

\item{permute}{should the interdistance matrix be permuted. Default is FALSE.}
}
\value{
A data.frame containing a P-value for the association.
}
\description{
Computes nominal P-value for the association between the genotype of a SNP and the splicing ratios of a gene.
}
\author{
Diego Garrido-Martín
}
\keyword{internal}
