% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD.filter.R
\name{LD.filter}
\alias{LD.filter}
\title{Linkage disequilibrium (LD) filter}
\usage{
LD.filter(genotype.gene, com.samples, tre.dist, th = 1, tol = 0.05,
  svQTL = FALSE)
}
\arguments{
\item{genotype.gene}{a data.frame with genotype info produced by 'read.bedix'.}

\item{com.samples}{a character vector with the common sample names between the genotype
and transcript relative expression files.}

\item{tre.dist}{a distance object from the transcript relative expression.}

\item{th}{r2 threshold over which SNPs will be merged. Default is 1.}

\item{svQTL.}{\code{LD.filter} only merges the SNP IDs if the correlation is higher than
\code{th} and the F scores for the eQTL test differ no more than a 5%. If svQTL = TRUE, 
it also requiresthat the F scores for the svQTL test differ no more than the 20%. Default is TRUE.}
}
\value{
a data.frame with columns:
\item{F}{the F score.}
\item{nb.groups}{the number of groups created by the genotypes.}
\item{md}{the maximum difference in splicing ratios between genotype groups.}
\item{tr.first, tr.second}{the two transcripts that change the most.}
\item{pv}{if \code{qform = TRUE} a P-value for the F score is computed}
\item{LD}{IDs of the SNPs that are in LD >= th, NA otherwise}
}
\description{
Groups SNPs according to LD.
Merges the SNP IDs of those variants that fulfill the following criteria:
\itemize{
\item{LD (r2) >= than a given user-defined threshold.}
\item{Relative differences in sQTL F score <= 0.05}
\item{(if svQTL = TRUE) Relative differences in sQTL F score <= 0.05}}
}
\author{
Diego Garrido-Martín
}
\keyword{internal}
