% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md.trans.R
\name{md.trans}
\alias{md.trans}
\title{MaxDiff splicing ratios computation}
\usage{
md.trans(sr.o, groups.o)
}
\arguments{
\item{sr.o}{a matrix with the transcript relative expression (samples x transcripts).}

\item{groups.o}{a factor with the genotype for each sample.}
}
\value{
A list with:
\item{md}{the maximum difference in splicing ratios between genotype groups.}
\item{tr.first, tr.second}{the two transcripts that change the most.}
}
\description{
Compute the maximum difference in transcript relative expression
between genotype groups.
}
\author{
Jean Monlong, Diego Garrido-Martín
}
\keyword{internal}
