% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.trans.exp.R
\name{nbDiffPt}
\alias{nbDiffPt}
\title{Number of non-identical splicing ratio points}
\usage{
nbDiffPt(sr)
}
\arguments{
\item{sr}{a data.frame with the splicing ratios (transcript x sample).}
}
\value{
the number of non-identical splicing ratios.
}
\description{
Compute the number of non-identical splicing ratios. Too few splicing ratios are
problematic for the permutation process. If many samples have exactly the same
splicing ratios, the permutation might create many identical permuted F scores.
}
\author{
Jean Monlong
}
\keyword{internal}
