% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.bedix.R
\name{read.bedix}
\alias{read.bedix}
\title{Retrieve a subset of an indexed BED file}
\usage{
read.bedix(file, subset.reg = NULL, header = TRUE)
}
\arguments{
\item{file}{a BED file containing the data to be read.
It should be compressed with \code{bgzip} and indexed by tabix algorithm.}

\item{subset.reg}{a data.frame or GRanges object with the regions to subset from.
If a data.frame, it must have columns named 'chr', 'start' and 'end'.}

\item{header}{should the first line of \code{file} be used as header. Default is \code{TRUE}.}
}
\value{
a data.frame with the retrieved BED information.
}
\description{
Using \code{Rsamtools} functions to manipulate tabix file, the indexed BED file is read
and the features present in the selected region are retrieved.
}
\author{
Jean Monlong, Diego Garrido-Martín
}
