% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.trans.exp.R
\name{relativize}
\alias{relativize}
\title{Relative expression computation}
\usage{
relativize(x, min.gene.exp = 0.01)
}
\arguments{
\item{x}{a vector of transcript expression.}

\item{min.gene.exp}{the minimum gene expression in the sample. If the
gene expression is too low it is safer to remove the sample from the
analysis.}
}
\value{
a vector with the relative expression.
}
\description{
Compute the relative expression of transcripts.
}
\author{
Jean Monlong
}
\keyword{internal}
