% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqtl.plot.R
\name{sqtl.plot}
\alias{sqtl.plot}
\title{Graphs for one sQTL}
\usage{
sqtl.plot(geneId, snpId, tre.df, genotype.f, gene.loc, genic.window = 5000,
  max.trans = 6, tr.str = NULL)
}
\arguments{
\item{geneId}{the gene ID}

\item{snpId}{the SNP ID}

\item{tre.df}{the data.frame with the transcript relative expression (produced by 'prepare.trans.exp').}

\item{genotype.f}{the path to the genotype file (should be compressed and indexed by 'index.genotype').}

\item{gene.loc}{a data.frame with the genes location. Columns 'chr', 'start',
'end' and 'geneId' are required.}

\item{genic.window}{the window(bp) around the gene in which the SNPs are tested. Default is 5000 (i.e. 5kb).}

\item{max.trans}{the maximum number of transcript shown. The least expressed transcript will be joined for visibility. Default is 6.}

\item{tr.str}{the data.frame with the structure of the transcripts. If NULL, not used.}
}
\value{
a list of ggplot graphs.
}
\description{
For a specific gene and SNP, this function produce the graph with the transcript relative expression in the different genotype groups. It also shows the structure of the transcripts, i.e. the CDS/UTR configuration.
}
\details{
The format of the transcript structure 'tr.str' should be : 'transId' the transcript ID, 'strand' the DNA strand, 'cdsStart' the start positions of the CDS regions separated by ",", 'cdsEnd' same with the end positions, 'urtStarts' and 'utrEnds' same with UTR regions. Look at the 'SplicingEventClassification' tutorial in the 'docs' folder of the github for some help on how to format such a file.
}
\author{
Jean Monlong
}
