% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqtl.seeker.p.R
\name{sqtl.seeker.p}
\alias{sqtl.seeker.p}
\title{permuted sQTL seeker}
\usage{
sqtl.seeker.p(tre.df, genotype.f, gene.loc, genic.window = 5000,
  min.nb.ext.scores = 100, nb.perm.max = 1000, verbose = TRUE)
}
\arguments{
\item{tre.df}{a data.frame with transcript relative expression
produced by 'prepare.trans.exp'. Same as in \code{sqtl.seeker}.}

\item{genotype.f}{the name of the genotype file. This file needs to
be ordered by position, compressed and indexed using 'index.genotype' or externally using tabix (samtools).
Must have column 'snpId'. Same as in \code{sqtl.seeker}.}

\item{gene.loc}{a data.frame with the genes location. Columns 'chr', 'start',
'end' and 'geneId' are required. Same as in \code{sqtl.seeker}.}

\item{genic.window}{the window(bp) around the gene in which the SNPs are tested. Default is 5000 (i.e. 5kb).
Same as in \code{sqtl.seeker}.}

\item{min.nb.ext.scores}{the minimum number of permuted  nominal P-values lower than
the lowest observed nominal P-value to allow the computation to stop. Default is 1e2.}

\item{nb.perm.max}{the maximum number of permutations. Default is 1e3.}

\item{verbose}{Default is TRUE. Mainly for debugging.}

\item{ncores}{if \code{ncores > 1} the parallel package is used to speed computation. Default is 1.
Note: Windows users must set ncores = 1 (see ... for more details)}
}
\value{
a data.frame with columns
\item{geneId}{the gene name}
\item{variants.cis}{the number of variants tested in cis}
\item{ld}{a LD estimate for the genomic window (median r2)}
\item{best.snp}{the SNP with the lowest nominal P-value}
\item{best.nominal.pv}{nominal P-value corresponding to the best SNP}
\item{shape1}{Beta distribution parameter shape1}
\item{shape2}{Beta distribution parameter shape2}
\item{nb.perms}{the number of permutations used for the empirical P-value computation}
\item{pv.emp}{empirical P-value based on permutations}
\item{pv.emp.beta}{empirical P-value based on the beta approximation}
\item{runtime}{computation time per gene}
}
\description{
\code{sqtl.seeker.p} description
}
\details{
Details
}
\author{
Diego Garrido-Martín
}
