% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqtls.p.R
\name{sqtls.p}
\alias{sqtls.p}
\title{sQTL retrieval from permuted phenotypes}
\usage{
sqtls.p(res.nominal.df, res.permuted.df, FDR = 0.05, method = "BH",
  md.min = 0.05, out.pdf = NULL, svQTL.removal = FALSE,
  FDR.svQTL = 0.05)
}
\arguments{
\item{res.nominal.df}{a data.frame, output of 'sqtl.seeker' with the nominal P-values
for each gene/SNP test.}

\item{res.permuted.df}{a data.frame, output of 'sqtl.seeker.p' with the empirical P-values
and the MLE estimates for the beta distribution parameters for each gene.}

\item{FDR}{the False Discovery Rate to call an association significant. Default is 0.05.}

\item{method}{the FDR approach. Either \code{"BH"} (Benjamini-Hochberg) or \code{"qvalue"} (Storey). The latter 
tends to be less stringent.  Default is "BH".}

\item{md.min}{the minimum MD (Maximum Difference) in relative expression required. Maximum
difference in relative expression (MD) gives an idea of the effect size of the association.
Default is 0.05.}

\item{out.pdf}{the name of the pdf file to create. If NULL (default), no pdf output
will be created. If non-NULL, the distribution of the P-values and a semi-volcano plot
(P-value vs MD) will be shown.}

\item{svQTL.removal}{if TRUE (and column 'pv.svQTL' is present in 'res.df') significant
sQTL which are also significant svQTLs are not reported. Default is FALSE.}

\item{FDR.svQTL}{the False Discovery Rate to call a svQTL, that may be removed from the final set of sQTLs. 
Note that svQTL FDR is computed on the pooled nominal P-values from the svQTL test.}
}
\description{
Retrieves sQTLs after two-level multiple testing correction and svQTL removal (if requested). The distribution of the P-values and 
a semi-volcano plot can be also displayed.
}
\details{
We consider two levels of multiple testing:
\itemize{
\item{1. Multiple genetic variants are tested per gene.}
\item{2. Multiple genes are tested genome-wide.}}
sQTLseekeR uses a permutation approach, implemented and detailed in \code{\link{sqtl.seeker.p}} to correct for
the former and false discovery rate (FDR) estimation to control for the latter. Both Benjamini-Hochberg (\code{\link{p.adjust}} method) 
and Storey (\code{\link{qvalue}} method) approaches for FDR can be applied.

If \code{svQTL.removal = TRUE} and svQTLs were tested in 'sqtl.seeker', gene/SNPs with
significant svQTL association (after multiple testing correction and similar FDR threshold)
are removed from the final set of significant sQTLs.
}
\author{
Diego Garrido-Martín
}
