\name{sQTLseekeR2-package}
\alias{sQTLseekeR2-package}
\alias{sQTLseekeR2}
\docType{package}
\title{Splicing QTL discovery tool}
\description{Association study between genotypic information and transcript relative expression from RNASeq, using a distance-based non-parametric MANOVA approach.}
\details{
\tabular{ll}{
Package: \tab sQTLseekeR2\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2017-20-03\cr
License: \tab GPL-2\cr
}
}
\author{Diego Garrido, Jean Monlong, Miquel Calvo, Ferran Reverter, Roderic Guigo.

Maintainer: Diego Garrido <dgarrimar@gmail.com>}
\references{Monlong, J. et al. Identification of genetic variants associated with alternative splicing using sQTLseekeR. Nat. Commun. 5:4698 doi: 10.1038/ncomms5698 (2014).}
\keyword{splicing, qtl, rna-seq, sequencing, transcriptome}
\examples{
}
