% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.trans.exp.R
\name{te.dispersion}
\alias{te.dispersion}
\title{Splicing dispersion computation}
\usage{
te.dispersion(tr)
}
\arguments{
\item{tr}{a data.frame with the splicing ratios (transcript x sample).}
}
\value{
a value for the dispersion.
}
\description{
Compute the dispersion from a gene's transcript expression matrix. 
Dispersion is computed as the mean Hellinger distance to the centroid.
}
\author{
Diego Garrido-Martín
}
\keyword{internal}
