% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compFscore.R
\name{compFscore}
\alias{compFscore}
\title{F score computation}
\usage{
compFscore(geno.df, tre.dist, tre.df, svQTL = FALSE, qform = TRUE)
}
\arguments{
\item{geno.df}{a data.frame of one row with the genotype information for each sample.}

\item{tre.dist}{a distance object from the transcript relative expression.}

\item{tre.df}{a data.frame with the transcript relative expression.}

\item{svQTL}{should svQTL test be performed in addition to sQTL. Default is FALSE.}

\item{qform}{should significance for the F score (sQTL test) be computed using 
the \code{\link[CompQuadForm]{davies}} method in the  \code{CompQuadForm} package?. 
Default is TRUE.}
}
\value{
a data.frame with columns:
\item{F}{the F score.}
\item{nb.groups}{the number of groups created by the genotypes.}
\item{md}{the maximum difference in splicing ratios between genotype groups.}
\item{tr.first, tr.second}{the two transcripts that change the most.}
\item{pv}{if \code{qform = TRUE} a P-value for the F score is computed}
}
\description{
Compute the F score, max diff ratio difference and transcripts that change the most.
Additionally, it can compute a P-value to assess the significance of the F score.
}
\author{
Diego Garrido-Martín, Jean Monlong
}
\keyword{internal}
