% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compPvalue.R
\name{compPvalue}
\alias{compPvalue}
\title{P-values computation.}
\usage{
compPvalue(
  res.df,
  tre.mt,
  min.nb.ext.scores = 1000,
  nb.perm.max = 1e+06,
  svQTL = FALSE
)
}
\arguments{
\item{res.df}{a data.frame with the F scores and number of groups.}

\item{tre.mt}{a matrix with the transcript relative expression (samples x transcripts).}

\item{min.nb.ext.scores}{the minimum number of permuted scores higher than
'F.lead' to allow the computation to stop. Default is 1000.}

\item{nb.perm.max}{the maximum number of permutations. Default is 1e6.}

\item{svQTL}{should svQTL test be performed instead of sQTL. Default is \code{FALSE}.}
}
\value{
An updated data.frame res.df with new columns 'pv' and 'nb.perms'.
}
\description{
Compute P-values from F scores using the methods available in the \code{\link{vegan}} package. 
Method \code{\link[vegan]{adonis}} is used for sQTL testing and method 
\code{\link[vegan]{betadisper}} for svQTLs.
}
\author{
Diego Garrido-Martín, Jean Monlong
}
\keyword{internal}
