% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/te.dispersion.R
\name{te.dispersion}
\alias{te.dispersion}
\title{Dispersion computation}
\usage{
te.dispersion(tr)
}
\arguments{
\item{tr}{a transcript expression matrix}
}
\value{
a value for the dispersion
}
\description{
Compute the dispersion from the transcript expression matrix. 
Dispersion is computed as the mean Hellinger distance to the centroid.
}
\author{
Diego Garrido-Martín
}
\keyword{internal}
